Reverse Night Mode function, behaves like a norma light switch
Night Mode switch renamed LED. 
Reverse switch function so that it behaves like a normal light switch. Either the Fan LED is ON or OFF. This helps with Siri Commands like "Turn of all the lights". The Fan LED light then get put into "Night Mode" or an OFF setting. This approach is far more logical within the HomeKit ecosystem.